requirements.txt
 # FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HubSpot SDK
hubspot-api-client==8.2.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
python-jose[cryptography]==3.3.0

# Environment and Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Logging
loguru==0.7.2

# CORS
fastapi-cors==0.0.6
